Extract MemberPage to its own file
This is step is needed in order to to run scraper_test against MemberPage
data.
