+++ scraper.rb
@@ -11,6 +11,8 @@ require 'pry'
 require 'open-uri/cached'
 OpenURI::Cache.cache_path = '.cache'
 
+require_rel 'lib'
+
 class String
   def tidy
     gsub(/[[:space:]]+/, ' ').strip
@@ -60,44 +62,6 @@ class MembersPage < Scraped::HTML
     def tds
       noko.css('td')
     end
-  end
-end
-
-class MemberPage < Scraped::HTML
-  decorator Scraped::Response::Decorator::AbsoluteUrls
-
-  field :id do
-    url.to_s.split('/')[-2]
-  end
-
-  field :name do
-    noko.css('p.person-name').text.tidy
-  end
-
-  field :faction do
-    noko.css('p.person-support').text.tidy
-  end
-
-  field :image do
-    noko.css('.person-img img/@src').text
-  end
-
-  field :term do
-    '6'
-  end
-
-  field :phone do
-    infotable.xpath('.//th[.="Телефон:"]//following-sibling::td').text
-  end
-
-  field :source do
-    url.to_s
-  end
-
-  private
-
-  def infotable
-    noko.css('table.person-inform-table')
   end
 end
